requirements.txt
falcon
pip
setuptools
sqlalchemy_i18n
phonenumbers
gunicorn
pymysql
jsonchema

